fix: Fixed play command

+++ commands/musica/play.js
@@ -4,7 +4,7 @@ module.exports = {
     desc: "Oir un cumbión",
     inVoiceChannel : true,
     run: async (client, message, args) => {
-        if(message.channel.id != 885939057144758327 || message.channel.id != 784173263227322389){
+        if(message.channel.id != 885939057144758327 && message.channel.id != 784173263227322389){
             message.delete(); 
             return message.channel.send(`Ponlo en el canal correcto n, no queremos spam ${message.author} pendejo `);
         }
